Automate switching between day/night color schemes
This sets up individual light and dark color scheme files for vim/neovim
and alacritty, adds a script (`set-color`) for placing them, and
includes two no-arg scripts for later keybinding (`day-mode` and
`night-mode`).

Thanks to this blog post for the idea to `cat` the YAML into place:
https://shuheikagawa.com/blog/2020/02/14/switching-colorschemes-of-vim-and-alacritty/

TIL YAML is concatenative!